requirements.txt
# LangChain and LangGraph dependencies
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-community==0.0.10

# HTTP client for API calls
aiohttp==3.9.1
requests==2.31.0

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Database and storage
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Vector database
chromadb==0.4.18
faiss-cpu==1.7.4

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0